Update GDAL merge script to organize source folder.
Update the gdal_merge script and add option to first scan the given
source_folder for new and unsorted tif files.

With this change the source folder will be scanned for folders with the day of
year as name (e.g 0012017 for day 1 of 2017) and process all tif files within.
The output file name is also no longer an option since the folder name will be
used for that.

To help with the organization of the folder a 'scan-folder' option was added to
automate generating new folders and moving the corresponding files in there.