Refactor filtering of band values.
Remove dedicated script to filter band values and move the logic to the TileFile
class. This class already read the entire band values for a file and the
filtering can happen here. Saves the step of reading and writing the data from
the projected file and speeds up the overall process.

The TileFile class also got scrubbed of unneeded calculations that changed with
copying and filtering the band data.