Separate out test helper methods to own file in conftest.py
Clean up the conftest.py and move the helper methods to a dedicated file. This
also removes the defined constants that now live in the constants file.

Added a utility class OutputFile to test_helper method that makes asserting on
the script output file easier having pre-parsed most of the information.